Add files via upload
This project aims to use Verilog to develop and verify a small RISC CPU. The CPU has a 32-bit
instruction size, 32 general-purpose registers, a program counter (PC), a return address control
stack, and a stack pointer (SP). Four instruction types are supported (R-type, I-type, J-type, and
S-type), and an ALU with a "zero" signal signifies a zero result. There are separate data and
instruction memory. A Datapath and a control path are included in the RTL design, and three
design alternatives are available: single-cycle, multi-cycle, and 5-stage pipelined processors, and
the team decided to build a multi-processor. Verification is carried out using a testbench and code
sequences executed in the designed processor.